Added the necessary code to deactivate the submit button on modifybatch.php.  Batches may now be checked in and checking is done to be sure all members of the batch have been checked in before the submit button is activated.  This is all working now.
Had to add a kludge variable to keep track of the batch ID number during the process.  I didn't think about that while I was developing all the Javascript to handle the processing of the members of the batch and didn't realize it until I was ready to write the AJAX function that made the backend check in call.

Still need to complete the AJAX function as well as the backend function that actually checks the batch in and makes the changes in the database.  Still a lot of work to be done but we're getting there.