admin.php - Created and tested the add user form which allows administrators to add new users.
configure.php - Changed the level of the modBatchLevel global variable.

ajax.php - Created and tested addNewUser().  It checks to see if the optional variables are set or not and sends a properly formatted object to the backend function.

main.js - Updated addUserCallback() and created and tested addUser().

main.php - Inserted code to handle the submit button click on the add user form.

This release of the code base contains full functionality.  At this point new features can be added and the code base cleaned up some.