Did some additional debugging and added a small feature.
functions.php - Added code to show the return date formatted as m/d/y.

mains.js - bug fix: added code to blank the variables that are used to track the batch that is being checked in.  If a user clicked the submit button again, without clicking on any of the other returning batch buttons, the batch that was just checked in would be checked in another time, one for each click of the button.  This code wipes the variables as soon as the check in process is completed which prevents the problem.

admin_setup.php - Added in a new-line, carriage return to clean up the output from the script.